+++ DSCResources/MSFT_ADDomainController/MSFT_ADDomainController.psm1
@@ -34,7 +34,9 @@ $script:localizedData = Get-LocalizedData -ResourceName 'MSFT_ADDomainController
         Provide the name of the site you want the Domain Controller to be added to.
 
     .PARAMETER InstallDns
-        Specifies if the Dns service will be installed or not. Default value is true.
+        Specifies if the DNS Server service should be installed and configured on the domain controller.
+        If this is not set, the default for the parameter InstallDns of the cmdlet [Install-ADDSDomainController]
+        (https://docs.microsoft.com/en-us/powershell/module/addsdeployment/install-addsdomaincontroller) is used
 #>
 function Get-TargetResource
 {
@@ -71,7 +73,7 @@ function Get-TargetResource
         $SiteName,
 
         [Parameter()]
-        [boolean]
+        [System.Boolean]
         $InstallDns
     )
 
@@ -86,6 +88,7 @@ function Get-TargetResource
         ReadOnlyReplica                     = $false
         AllowPasswordReplicationAccountName = $null
         DenyPasswordReplicationAccountName  = $null
+        InstallDns                          = $InstallDNs
     }
 
     Write-Verbose -Message (
@@ -187,7 +190,9 @@ function Get-TargetResource
         Provides a list of the users, computers, and groups to add to the password replication denied list.
 
     .PARAMETER InstallDns
-        Specifies if the Dns service will be installed or not. Default value is true.
+        Specifies if the DNS Server service should be installed and configured on the domain controller.
+        If this is not set, the default for the parameter InstallDns of the cmdlet [Install-ADDSDomainController]
+        (https://docs.microsoft.com/en-us/powershell/module/addsdeployment/install-addsdomaincontroller) is used
 #>
 function Set-TargetResource
 {
@@ -256,7 +261,7 @@ function Set-TargetResource
         $DenyPasswordReplicationAccountName,
 
         [Parameter()]
-        [boolean]
+        [System.Boolean]
         $InstallDns
     )
 
@@ -328,9 +333,9 @@ function Set-TargetResource
             $installADDSDomainControllerParameters.Add('NoGlobalCatalog', $true)
         }
 
-        if ($PSBoundParameters.ContainsKey('InstallDns') -and $InstallDns -eq $false)
+        if ($PSBoundParameters.ContainsKey('InstallDns'))
         {
-            $installADDSDomainControllerParameters.Add('InstallDns', $false)
+            $installADDSDomainControllerParameters.Add('InstallDns', $InstallDns)
         }
 
         if (-not [System.String]::IsNullOrWhiteSpace($InstallationMediaPath))
@@ -534,7 +539,10 @@ function Set-TargetResource
         Provides a list of the users, computers, and groups to add to the password replication denied list.
 
     .PARAMETER InstallDns
-        Specifies if the Dns service will be installed or not. Default value is true.
+        Specifies if the DNS Server service should be installed and configured on the domain controller.
+        If this is not set, the default for the parameter InstallDns of the cmdlet [Install-ADDSDomainController]
+        (https://docs.microsoft.com/en-us/powershell/module/addsdeployment/install-addsdomaincontroller) is used
+        Not used in Test-TargetResource.
 #>
 function Test-TargetResource
 {
@@ -593,7 +601,7 @@ function Test-TargetResource
         $DenyPasswordReplicationAccountName,
 
         [Parameter()]
-        [boolean]
+        [System.Boolean]
         $InstallDns
     )